+++ pub/flutter_unrouter/lib/src/core/route_definition.dart
@@ -1,5 +1,3 @@
-import 'dart:async';
-
 import 'package:flutter/widgets.dart';
 import 'package:unrouter/unrouter.dart'
     hide